refactor: move list_qualities_and_formats to cli module
